object: use fopen on windows
on Windows, fdopen is broken.
use fopen instead.

Fixes #21.
